+++ internal/internal.go
@@ -1,0 +1,1 @@
+package internal
